\name{print.lmdt}
\alias{print.lmdt}
\alias{summary.lmdt}
\title{Methods for LMDT Test Results}
\usage{
\method{print}{lmdt}(x, digits = 4, ...)

\method{summary}{lmdt}(object, ...)
}
\arguments{
\item{x}{An object of class \code{"lmdt"}.}

\item{digits}{Number of significant digits.}

\item{object}{An object of class \code{"lmdt"}.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns the input object.
}
\description{
Print and summary methods for objects of class \code{"lmdt"}.
}
